requirements.txt
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.3
Jinja2==3.1.2
loguru==0.6.0
MarkupSafe==2.1.1
paramiko==2.11.1
progressbar==2.5
pycparser==2.21
pydantic==1.10.2
PyNaCl==1.5.0
PyYAML==6.0
scp==0.14.4
six==1.16.0
typer==0.6.1
typing_extensions==4.3.0
win32-setctime==1.1.0
